move SettlementError into single_order_solver
